% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{deriv_dirichlet_process}
\alias{deriv_dirichlet_process}
\title{Derivative of the mean of the Dirichlet process distribution}
\usage{
deriv_dirichlet_process(alpha, n)
}
\arguments{
\item{alpha}{Concentration parameter (positive scalar or vector).}

\item{n}{Sample size (positive integer or vector).}
}
\description{
Derivative of the mean of the Dirichlet process distribution
}
