% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{hubbell}
\alias{hubbell}
\title{Hubbell family of generalized linear models}
\usage{
hubbell(link = "polyseries", sigma = 0)
}
\arguments{
\item{link}{The link function to use. Available options are \code{dp}, and \code{polyseries}}

\item{sigma}{Hyperparameter determining the polynomial growth. Must be \code{sigma < 1} for
link  \code{link = 'polyseries'}. It is set to 0 when \code{link = 'dp'} (logarithmic growth).}
}
\description{
Hubbell family of generalized linear models
}
