% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HubbellGLM.R
\name{HubbellGLM}
\alias{HubbellGLM}
\title{Hubbell regression
This function mimics the GLM structure. We do so to avoid conflicts with the
current glm implementation (if it gets updated)}
\usage{
HubbellGLM(
  formula,
  family = hubbell(sigma = 0),
  data,
  weights,
  subset,
  na.action,
  start = NULL,
  etastart,
  mustart,
  offset,
  control = list(...),
  model = TRUE,
  x = FALSE,
  y = TRUE,
  singular.ok = TRUE,
  contrasts = NULL,
  ...
)
}
\arguments{
\item{formula}{Same as \code{stats::glm}}

\item{family}{a member of the hubbell family for \code{glm.fit.hubbell}.}

\item{data}{Same as \code{stats::glm}}

\item{weights}{Same as \code{stats::glm}}

\item{subset}{Same as \code{stats::glm}}

\item{na.action}{Same as \code{stats::glm}}

\item{start}{Same as \code{stats::glm}}

\item{etastart}{Same as \code{stats::glm}}

\item{mustart}{Same as \code{stats::glm}}

\item{offset}{Same as \code{stats::glm}}

\item{control}{Same as \code{stats::glm}}

\item{model}{Same as \code{stats::glm}}

\item{x}{Same as \code{stats::glm}}

\item{y}{Same as \code{stats::glm}}

\item{singular.ok}{Same as \code{stats::glm}}

\item{contrasts}{Same as \code{stats::glm}}

\item{...Same}{as \code{stats::glm}}
}
\value{
An object of class \code{c('HubbellGLM', 'lm')}, which inherits from the class "lm"
}
\description{
Hubbell regression
This function mimics the GLM structure. We do so to avoid conflicts with the
current glm implementation (if it gets updated)
}
