% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{inv_polyseries}
\alias{inv_polyseries}
\title{Function to invert the polyseries}
\usage{
inv_polyseries(mu_target, size, sigma)
}
\arguments{
\item{mu_target}{Target mean}

\item{size}{Number of units in the sequence}

\item{sigma}{Parameter for the growth. sigma = 0 is the logarithmic growth}
}
\description{
Function to invert the polyseries
}
