% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{make.link.hubbell}
\alias{make.link.hubbell}
\title{Link function for the Hubbell generalized linear model}
\usage{
make.link.hubbell(link)
}
\arguments{
\item{link}{Name for the link function}
}
\description{
Link function for the Hubbell generalized linear model
}
