% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link.R
\name{mean_dirichlet_process}
\alias{mean_dirichlet_process}
\title{Mean of the n. of species from Dirichlet process}
\usage{
mean_dirichlet_process(alpha, n)
}
\arguments{
\item{alpha}{parameter of the Dirichlet process}

\item{n}{Number of units}
}
\description{
Mean of the n. of species from Dirichlet process
}
