% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.HubbellGLM}
\alias{predict.HubbellGLM}
\title{Function to predict the response from a HubbellGLM model.}
\usage{
\method{predict}{HubbellGLM}(
  object,
  newdata = NULL,
  type = c("link", "response", "terms"),
  se.fit = FALSE,
  dispersion = NULL,
  terms = NULL,
  na.action = na.pass,
  ...
)
}
\description{
Function to predict the response from a HubbellGLM model.
}
\keyword{internal}
