% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.HubbellGLM}
\alias{summary.HubbellGLM}
\title{Summary function for the HubbellGLM method}
\usage{
\method{summary}{HubbellGLM}(
  object,
  dispersion = NULL,
  correlation = FALSE,
  symbolic.cor = FALSE,
  ...
)
}
\description{
Summary function for the HubbellGLM method
}
